encapsulate text watcher
- use same space trimming logic for search fields
